+++ python/fate/arch/dataframe/io/_json_serialization.py
@@ -12,11 +12,10 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from fate.arch.context.io.data import df
 
 from .._dataframe import DataFrame
-from ._json_schema import build_schema, parse_schema
 from ..manager import DataManager
+from ._json_schema import build_schema, parse_schema
 
 
 def _serialize(ctx, data):
@@ -67,15 +66,13 @@ def _serialize(ctx, data):
     schema = build_schema(data)
 
     from ..ops._transformer import transform_block_to_list
+
     serialize_data = transform_block_to_list(data.block_table, data.data_manager)
     serialize_data.schema = schema
     return serialize_data
 
 
 def serialize(ctx, data):
-    if isinstance(data, df.Dataframe):
-        data = data.data
-
     return _serialize(ctx, data)
