+++ python/fate/components/core/__init__.py
@@ -1,13 +1,10 @@
 from . import _cpn_reexport as cpn
 from ._cpn_search import list_components, load_component
-from ._label import T_LABEL
 from ._load_computing import load_computing
 from ._load_device import load_device
 from ._load_federation import load_federation
-from ._role import ARBITER, GUEST, HOST, LOCAL, T_ROLE, Role, load_role
-from ._stage import T_STAGE, Stage, load_stage
-from .component_desc._component import Component
-from .component_desc._component_io import ComponentExecutionIO
+from .component_desc import Component, ComponentExecutionIO
+from .essential import ARBITER, GUEST, HOST, LABELS, LOCAL, Role, Stage
 
 __all__ = [
     "Component",
@@ -18,14 +15,10 @@ __all__ = [
     "load_device",
     "load_computing",
     "load_federation",
-    "load_role",
-    "load_stage",
     "Role",
     "Stage",
-    "T_ROLE",
-    "T_STAGE",
-    "T_LABEL",
     "ARBITER",
     "GUEST",
     "HOST",
+    "LOCAL",
 ]
