+++ python/fate/components/core/component_desc/_component.py
@@ -53,9 +53,16 @@ import inspect
 import logging
 from typing import List, Optional
 
-from .._role import T_ROLE, Role
-from .._stage import CROSS_VALIDATION, DEFAULT, PREDICT, TRAIN
-from ._artifact_base import ComponentArtifactDescribes
+from fate.components.core.essential import (
+    CROSS_VALIDATION,
+    DEFAULT,
+    PREDICT,
+    TRAIN,
+    Role,
+    Stage,
+)
+
+from ._component_artifact import ComponentArtifactDescribes
 from ._parameter import ComponentParameterDescribes
 
 logger = logging.getLogger(__name__)
@@ -65,7 +72,7 @@ class Component:
     def __init__(
         self,
         name: str,
-        roles: List[T_ROLE],
+        roles: List[Role],
         provider,
         version,
         description,
@@ -274,9 +281,9 @@ def _component(name, roles, provider, version, description, is_subcomponent):
             artifacts = ComponentArtifactDescribes()
 
         if is_subcomponent:
-            artifacts.set_stages([cpn_name])
+            artifacts.set_stages([Stage.from_str(cpn_name)])
         else:
-            artifacts.set_stages([DEFAULT.name])
+            artifacts.set_stages([DEFAULT])
         desc = description
         if desc is None:
             desc = inspect.getdoc(f)
