+++ python/fate/components/core/component_desc/artifacts/__init__.py
@@ -1,0 +1,49 @@
+from ._base_type import ArtifactDescribe, ArtifactType
+from .data import (
+    data_directory_input,
+    data_directory_inputs,
+    data_directory_output,
+    data_directory_outputs,
+    dataframe_input,
+    dataframe_inputs,
+    dataframe_output,
+    dataframe_outputs,
+    table_input,
+    table_inputs,
+)
+from .metric import json_metric_output, json_metric_outputs
+from .model import (
+    json_model_input,
+    json_model_inputs,
+    json_model_output,
+    json_model_outputs,
+    model_directory_input,
+    model_directory_inputs,
+    model_directory_output,
+    model_directory_outputs,
+)
+
+__all__ = [
+    "ArtifactType",
+    "ArtifactDescribe",
+    "json_model_input",
+    "json_model_inputs",
+    "json_model_output",
+    "json_model_outputs",
+    "model_directory_input",
+    "model_directory_inputs",
+    "model_directory_output",
+    "model_directory_outputs",
+    "dataframe_input",
+    "dataframe_inputs",
+    "dataframe_output",
+    "dataframe_outputs",
+    "table_input",
+    "table_inputs",
+    "data_directory_input",
+    "data_directory_inputs",
+    "data_directory_output",
+    "data_directory_outputs",
+    "json_metric_output",
+    "json_metric_outputs",
+]
