+++ python/fate/components/core/essential/__init__.py
@@ -1,0 +1,3 @@
+from ._label import LABELS
+from ._role import ARBITER, GUEST, HOST, LOCAL, Role
+from ._stage import CROSS_VALIDATION, DEFAULT, PREDICT, TRAIN, Stage
