+++ python/fate/components/core/essential/_label.py
@@ -12,14 +12,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-class Dataframe:
-    def __init__(self, frames, num_features, num_samples) -> None:
-        self.data = frames
-        self.num_features = num_features
-        self.num_samples = num_samples
 
-    def __len__(self):
-        return self.num_samples
 
-    def to_local(self):
-        return self.data.to_local()
+class LABELS:
+    TRAINABLE = "trainable"
